refactor: extract MenuBarLevelMeter.swift from MenuBarView.swift
Extracts LevelMeterView (View, 36 lines) + MicrophoneInputLevelMonitor
(@MainActor ObservableObject, 140 lines) from MenuBarView.swift into a new
MenuBarLevelMeter.swift.

Byte-identical move; SHA-256 00900bac93655e45974c8b741d6aa5ac399c1e5e28dd2c9f58bdfb008e090a63
verified against origin/main:MenuBarView.swift lines 700-877.

Imports: AVFoundation + CoreGraphics + Combine + SwiftUI.

Reduces MenuBarView.swift from 877 to 698 lines.

Note: This is a SwiftUI view/observable extraction. The extracted types (view
+ observable) are self-contained. LevelMeterView is used by MenuBarView; the
call site is name-based (not location-based). SwiftUI observation preserved.

⚠️ Manual smoke check recommended before merge: render the menu bar and
verify the microphone level meter animates identically to pre-move.

Tests: 667 (unchanged).
